huge refactor: re-integrate main encoding / decoding logi back into taskmodule; construct taskmodule in the model and use that to get the metric
