+++ src/pie_modules/taskmodules/pointer_network_taskmodule.py
@@ -1,10 +1,24 @@
 import dataclasses
 import logging
-from typing import Any, Dict, Iterator, List, Optional, Sequence, Tuple, Type, Union
+from collections import Counter, defaultdict
+from functools import cmp_to_key
+from typing import (
+    Any,
+    Dict,
+    Iterator,
+    List,
+    Optional,
+    Sequence,
+    Set,
+    Tuple,
+    Type,
+    Union,
+)
 
+import numpy as np
 import torch
 from pytorch_ie import AnnotationLayer, Document
-from pytorch_ie.annotations import Span
+from pytorch_ie.annotations import BinaryRelation, LabeledSpan, Span
 from pytorch_ie.core import Annotation, TaskEncoding, TaskModule
 from pytorch_ie.core.taskmodule import (
     InputEncoding,
@@ -13,6 +27,7 @@ from pytorch_ie.core.taskmodule import (
     TaskBatchEncoding,
 )
 from pytorch_ie.documents import TextBasedDocument, TokenBasedDocument
+from torchmetrics import Metric
 from transformers import AutoTokenizer, PreTrainedTokenizer
 from typing_extensions import TypeAlias
 
@@ -23,11 +38,13 @@ from pie_modules.documents import (
 
 from ..document.processing import token_based_document_to_text_based, tokenize_document
 from ..utils import resolve_type
-from .components.common import BatchableMixin
-from .components.pointer_network import (
-    EncodingWithIdsAndOptionalCpmTag,
-    PointerNetworkSpanAndRelationEncoderDecoder,
+from .common.mixins import BatchableMixin, HasBuildMetric, HasDecodeAnnotations
+from .pointer_network.annotation_encoder_decoder import (
+    BinaryRelationEncoderDecoder,
+    LabeledSpanEncoderDecoder,
+    SpanEncoderDecoderWithOffset,
 )
+from .pointer_network.metrics import AnnotationLayerMetric
 
 logger = logging.getLogger(__name__)
 
@@ -45,13 +62,27 @@ class InputEncodingType(BatchableMixin):
         return len(self.src_tokens)
 
 
+@dataclasses.dataclass
+class EncodingWithIdsAndOptionalCpmTag(BatchableMixin):
+    tgt_tokens: List[int]
+    CPM_tag: Optional[List[List[int]]] = None
+
+    @property
+    def tgt_seq_len(self) -> int:
+        return len(self.tgt_tokens)
+
+    @property
+    def tgt_attention_mask(self) -> List[int]:
+        return [1] * len(self.tgt_tokens)
+
+
 TargetEncodingType: TypeAlias = EncodingWithIdsAndOptionalCpmTag
 TaskEncodingType: TypeAlias = TaskEncoding[
     DocumentType,
     InputEncodingType,
     TargetEncodingType,
 ]
-TaskOutput: TypeAlias = EncodingWithIdsAndOptionalCpmTag
+TaskOutputType: TypeAlias = EncodingWithIdsAndOptionalCpmTag
 
 
 def _span_is_in_partition(span: Span, partition: Optional[Span] = None):
@@ -63,28 +94,49 @@ def _span_is_in_partition(span: Span, partition: Optional[Span] = None):
     )
 
 
+def cmp_src_rel(v1: BinaryRelation, v2: BinaryRelation) -> int:
+    if not all(isinstance(ann, LabeledSpan) for ann in [v1.head, v1.tail, v2.head, v2.tail]):
+        raise Exception(f"expected LabeledSpan, but got: {v1}, {v2}")
+    if v1.head.start == v2.head.start:  # v1[0]["from"] == v2[0]["from"]:
+        return v1.tail.start - v2.tail.start  # v1[1]["from"] - v2[1]["from"]
+    return v1.head.start - v2.head.start  # v1[0]["from"] - v2[0]["from"]
+
+
 # TODO: use enable BucketSampler (just mentioning here because no better place available for now)
 # see https://github.com/Lightning-AI/lightning/pull/13640#issuecomment-1199032224
 
 
 @TaskModule.register()
 class PointerNetworkTaskModule(
+    HasBuildMetric,
+    HasDecodeAnnotations[TaskOutputType],
     TaskModule[
         DocumentType,
         InputEncoding,
         TargetEncoding,
         TaskBatchEncoding,
         ModelBatchOutput,
-        TaskOutput,
-    ]
+        TaskOutputType,
+    ],
 ):
-    # Note that we overwrite self.is_prepared with self.annotation_encoder_decoder.is_prepared, so it is fine
-    # to initialize the taskmodule *with* a value for annotation_encoder_decoder_kwargs, but _prepare(documents)
-    # gets still called when the taskmodule is used in a pipeline.
-    PREPARED_ATTRIBUTES = ["annotation_encoder_decoder_kwargs"]
+    PREPARED_ATTRIBUTES = ["labels_per_layer"]
 
     def __init__(
         self,
+        # specific for this use case
+        span_layer_name: str = "labeled_spans",
+        relation_layer_name: str = "binary_relations",
+        none_label: str = "none",
+        loop_dummy_relation_name: str = "loop",
+        ignore_error_types: Optional[List[str]] = None,
+        # generic pointer network
+        label_tokens: Optional[Dict[str, str]] = None,
+        label_representations: Optional[Dict[str, str]] = None,
+        labels_per_layer: Optional[Dict[str, List[str]]] = None,
+        exclude_labels_per_layer: Optional[Dict[str, List[str]]] = None,
+        # target encoding
+        max_target_length: Optional[int] = None,
+        create_constraints: bool = False,
         # tokenization
         document_type: str = "pytorch_ie.documents.TextDocumentWithLabeledSpansBinaryRelationsAndLabeledPartitions",
         tokenized_document_type: str = "pie_modules.documents.TokenDocumentWithLabeledSpansBinaryRelationsAndLabeledPartitions",
@@ -93,14 +145,6 @@ class PointerNetworkTaskModule(
         tokenizer_kwargs: Optional[Dict[str, Any]] = None,
         partition_layer_name: Optional[str] = None,
         annotation_field_mapping: Optional[Dict[str, str]] = None,
-        # how to encode and decode the annotations
-        annotation_encoder_decoder_name: str = "pointer_network_span_and_relation",
-        annotation_encoder_decoder_kwargs: Optional[Dict[str, Any]] = None,
-        label_tokens: Optional[Dict[str, str]] = None,
-        label_representations: Optional[Dict[str, str]] = None,
-        # target encoding
-        max_target_length: Optional[int] = None,
-        create_constraints: bool = False,
         # logging
         log_first_n_examples: Optional[int] = None,
         **kwargs,
@@ -124,30 +168,30 @@ class PointerNetworkTaskModule(
         self.annotation_field_mapping = annotation_field_mapping or dict()
         self.partition_layer_name = partition_layer_name
 
+        # for this specific use case: end-to-end relation extraction
+        self.span_layer_name = span_layer_name
+        self.relation_layer_name = relation_layer_name
+        self.none_label = none_label
+        self.loop_dummy_relation_name = loop_dummy_relation_name
+        self.ignore_error_types = ignore_error_types or []
+        # will be set in _post_prepare()
+        self.relation_encoder_decoder: BinaryRelationEncoderDecoder
+
+        # collected in prepare(), if not passed in
+        self.labels_per_layer = labels_per_layer
+        self.exclude_labels_per_layer = exclude_labels_per_layer or {}
+
         # how to encode and decode the annotations
-        self.annotation_encoder_decoder_name = annotation_encoder_decoder_name
-        self.annotation_encoder_decoder_kwargs = annotation_encoder_decoder_kwargs or {}
-        if self.annotation_encoder_decoder_name == "pointer_network_span_and_relation":
-            # do not pass bos_token and eos_token directly as constructor arguments,
-            # because they may be already in annotation_encoder_decoder_kwargs
-            self.annotation_encoder_decoder_kwargs["bos_token"] = self.tokenizer.bos_token
-            self.annotation_encoder_decoder_kwargs["eos_token"] = self.tokenizer.eos_token
-            self.annotation_encoder_decoder = PointerNetworkSpanAndRelationEncoderDecoder(
-                create_constraints=create_constraints,
-                max_length=max_target_length,
-                **self.annotation_encoder_decoder_kwargs,
-            )
-        else:
-            raise Exception(
-                f"unknown annotation_encoder_decoder_name: {self.annotation_encoder_decoder_name}"
-            )
+        self.bos_token = self.tokenizer.bos_token
+        self.eos_token = self.tokenizer.eos_token
         self.label_tokens = label_tokens or dict()
         self.label_representations = label_representations or dict()
 
         # target encoding
         self.max_target_length = max_target_length
+        self.create_constraints = create_constraints
         self.pad_values = {
-            "tgt_tokens": self.annotation_encoder_decoder.target_pad_id,
+            "tgt_tokens": self.target_pad_id,
             "tgt_attention_mask": 0,
             "src_tokens": self.tokenizer.pad_token_id,
             "src_attention_mask": 0,
@@ -175,11 +219,44 @@ class PointerNetworkTaskModule(
         return self._tokenized_document_type
 
     @property
-    def is_prepared(self):
-        return self.annotation_encoder_decoder.is_prepared
+    def layer_names(self) -> List[str]:
+        return [self.span_layer_name, self.relation_layer_name]
 
-    def _prepare(self, documents: Sequence[DocumentType]):
-        self.annotation_encoder_decoder._prepare(documents=documents)
+    @property
+    def special_targets(self) -> list[str]:
+        return [self.bos_token, self.eos_token]
+
+    @property
+    def special_target2id(self) -> Dict[str, int]:
+        return {target: idx for idx, target in enumerate(self.special_targets)}
+
+    @property
+    def target_pad_id(self) -> int:
+        return self.special_target2id[self.eos_token]
+
+    @property
+    def generation_kwargs(self) -> Dict[str, Any]:
+        return {
+            "no_repeat_ngram_size": 7,
+            # TODO: add this when it looks really solid (currently strange behavior)
+            # "prefix_allowed_tokens_fn": self._prefix_allowed_tokens_fn,
+        }
+
+    def _prepare(self, documents: Sequence[DocumentType]) -> None:
+        # collect all labels
+        labels: Dict[str, Set[str]] = {layer_name: set() for layer_name in self.layer_names}
+        for doc in documents:
+            for layer_name in self.layer_names:
+                exclude_labels = self.exclude_labels_per_layer.get(layer_name, [])
+                labels[layer_name].update(
+                    ac.label for ac in doc[layer_name] if ac.label not in exclude_labels
+                )
+
+        self.labels_per_layer = {
+            # sort labels to ensure deterministic order
+            layer_name: sorted(labels)
+            for layer_name, labels in labels.items()
+        }
 
     def construct_label_token(self, label: str) -> str:
         return self.label_tokens.get(label, f"<<{label}>>")
@@ -187,19 +264,54 @@ class PointerNetworkTaskModule(
     def get_label_representation(self, label: str) -> str:
         return self.label_representations.get(label, label)
 
-    def _post_prepare(self):
-        self.annotation_encoder_decoder._post_prepare()
-        # This is a bit hacky, but we need to update the kwargs of the annotation-encoder-decoder with its
-        # prepared attributes, because they are required when instantiating it from config, i.e. when
-        # taskmodule._prepare(documents) is *not* called.
-        self.annotation_encoder_decoder_kwargs.update(
-            self.annotation_encoder_decoder.prepared_attributes
+    def _post_prepare(self) -> None:
+        # set up labels
+        if self.labels_per_layer is None:
+            raise Exception("labels_per_layer is not defined. Call prepare() first or pass it in.")
+        self.labels: List[str] = [self.none_label]
+        for layer_name in self.layer_names:
+            self.labels.extend(self.labels_per_layer[layer_name])
+        if len(set(self.labels)) != len(self.labels):
+            raise Exception(f"labels are not unique: {self.labels}")
+
+        # set up targets and ids
+        self.targets: List[str] = self.special_targets + self.labels
+        self.target2id: Dict[str, int] = {target: idx for idx, target in enumerate(self.targets)}
+
+        # generic ids
+        self.eos_id: int = self.target2id[self.eos_token]
+        self.bos_id: int = self.target2id[self.bos_token]
+
+        # span and relation ids
+        self.span_ids: List[int] = [
+            self.target2id[label] for label in self.labels_per_layer[self.span_layer_name]
+        ]
+        self.relation_ids: List[int] = [
+            self.target2id[label] for label in self.labels_per_layer[self.relation_layer_name]
+        ]
+        # the none id is used for the dummy relation which models out-of-relation spans
+        self.none_id: int = self.target2id[self.none_label]
+
+        # helpers (same as targets / target2id, but only for labels)
+        self.label2id: Dict[str, int] = {label: self.target2id[label] for label in self.labels}
+        self.id2label: Dict[int, str] = {idx: label for label, idx in self.label2id.items()}
+        self.label_ids: List[int] = [self.label2id[label] for label in self.labels]
+
+        # annotation-encoder-decoders
+        span_encoder_decoder = SpanEncoderDecoderWithOffset(offset=self.pointer_offset)
+        labeled_span_encoder_decoder = LabeledSpanEncoderDecoder(
+            span_encoder_decoder=span_encoder_decoder,
+            label2id=self.label2id,
+        )
+        self.relation_encoder_decoder = BinaryRelationEncoderDecoder(
+            head_encoder_decoder=labeled_span_encoder_decoder,
+            tail_encoder_decoder=labeled_span_encoder_decoder,
+            label2id=self.label2id,
+            loop_dummy_relation_name=self.loop_dummy_relation_name,
+            none_label=self.none_label,
         )
 
-        label2token = {
-            label: self.construct_label_token(label=label)
-            for label in self.annotation_encoder_decoder.labels
-        }
+        label2token = {label: self.construct_label_token(label=label) for label in self.labels}
         if len(set(label2token.values())) != len(label2token):
             raise Exception(f"label2token values are not unique: {label2token}")
 
@@ -220,8 +332,8 @@ class PointerNetworkTaskModule(
         )
 
         # target tokens are the special tokens plus the mapped label tokens
-        self.target_tokens: List[str] = self.annotation_encoder_decoder.special_targets + [
-            label2token[label] for label in self.annotation_encoder_decoder.labels
+        self.target_tokens: List[str] = self.special_targets + [
+            label2token[label] for label in self.labels
         ]
         self.target_token_ids: List[int] = self.tokenizer.convert_tokens_to_ids(self.target_tokens)
 
@@ -248,6 +360,273 @@ class PointerNetworkTaskModule(
                 )
             self.label_embedding_weight_mapping[label_token_idx] = source_indices
 
+    @property
+    def pointer_offset(self) -> int:
+        return len(self.targets)
+
+    @property
+    def target_ids(self) -> Set[int]:
+        return set(range(self.pointer_offset))
+
+    def build_metric(self, stage: Optional[str] = None) -> Metric:
+        return AnnotationLayerMetric(
+            eos_id=self.eos_id,
+            taskmodule=self,
+            layer_names=self.layer_names,
+            decode_annotations_func=self.decode_annotations,
+        )
+
+    def sanitize_sequence(
+        self,
+        tag_seq: List[int],
+    ) -> Tuple[List[Tuple[int, int, int, int, int, int, int]], Dict[str, int]]:
+        # TODO: count total amounts instead of returning bool values.
+        #  This requires to also count "total" (maybe also "skipped" and "correct").
+        invalid = {
+            "len": 0,
+            "order": 0,
+            "cross": 0,
+            "cover": 0,
+        }  # , "total": 0 , "skipped": 0, "correct": 0}
+        skip = False
+        pairs = []
+        cur_pair: List[int] = []
+        if len(tag_seq):
+            for i in tag_seq:
+                if i in self.relation_ids or (i == self.none_id and len(cur_pair) == 6):
+                    cur_pair.append(i)
+                    if len(cur_pair) != 7:
+                        skip = True
+                        invalid["len"] = 1
+                    elif self.none_id in cur_pair:
+                        # tag = set([cur_pair[2], cur_pair[5], cur_pair[6]])
+                        if not (
+                            cur_pair[2] in self.target_ids
+                            and cur_pair[5] in self.target_ids
+                            and cur_pair[6] in self.target_ids
+                        ):
+                            # if not tag.issubset(add_token):
+                            skip = True
+                        else:
+                            skip = False
+                    else:  # The decoding length is correct (解码长度正确)
+                        # Check for correct position (检查位置是否正确) <s1,e1,t1,s2,e2,t2,t3>
+                        if cur_pair[0] > cur_pair[1] or cur_pair[3] > cur_pair[4]:
+                            if "cover" not in self.ignore_error_types:
+                                skip = True
+                            invalid["order"] = 1
+                        elif not (cur_pair[1] < cur_pair[3] or cur_pair[0] > cur_pair[4]):
+                            skip = True
+                            invalid["cover"] = 1
+                        if (
+                            cur_pair[2] in self.relation_ids
+                            or cur_pair[5] in self.relation_ids
+                            or cur_pair[6] in self.span_ids
+                        ):
+                            # Consider making an additional layer of restrictions to prevent misalignment
+                            # of the relationship and span tags (可以考虑做多一层限制，防止relation 和 span标签错位)
+                            if "cross" not in self.ignore_error_types:
+                                skip = True
+                            invalid["cross"] = 1
+                        # tag = set([cur_pair[2], cur_pair[5], cur_pair[6]])
+                        RC_idx = self.relation_ids + self.span_ids
+                        if not (
+                            cur_pair[2] in RC_idx
+                            and cur_pair[5] in RC_idx
+                            and cur_pair[6] in RC_idx
+                        ):
+                            # if not tag.issubset(self.relation_idx+self.span_idx):
+                            skip = True
+                            invalid["cross"] = 1
+
+                    if skip:
+                        skip = False
+                        # invalid["skipped"] += 1
+                    else:
+                        if len(cur_pair) != 7:
+                            raise Exception(f"expected 7 entries, but got: {cur_pair}")
+                        pairs.append(tuple(cur_pair))
+                        # invalid["correct"] += 1
+                    cur_pair = []
+                else:
+                    cur_pair.append(i)
+
+        # invalid["total"] = invalid["correct"] + invalid["skipped"]
+
+        # ignore type because of tuple length
+        return pairs, invalid  # type: ignore
+
+    def encode_annotations(
+        self, layers: Dict[str, List[Annotation]], metadata: Optional[Dict[str, Any]] = None
+    ) -> TaskOutputType:
+        if not set(layers.keys()) == set(self.layer_names):
+            raise Exception(f"unexpected layers: {layers.keys()}. expected: {self.layer_names}")
+
+        if self.labels_per_layer is None:
+            raise Exception("labels_per_layer is not defined. Call prepare() first or pass it in.")
+
+        # encode relations
+        all_relation_arguments = set()
+        relation_encodings = dict()
+        for rel in layers[self.relation_layer_name]:
+            if not isinstance(rel, BinaryRelation):
+                raise Exception(f"expected BinaryRelation, but got: {rel}")
+            if rel.label in self.labels_per_layer[self.relation_layer_name]:
+                encoded_relation = self.relation_encoder_decoder.encode(
+                    annotation=rel, metadata=metadata
+                )
+                if encoded_relation is None:
+                    raise Exception(f"failed to encode relation: {rel}")
+                relation_encodings[rel] = encoded_relation
+                all_relation_arguments.update([rel.head, rel.tail])
+
+        # encode spans that are not arguments of any relation
+        no_relation_spans = [
+            span for span in layers[self.span_layer_name] if span not in all_relation_arguments
+        ]
+        for span in no_relation_spans:
+            dummy_relation = BinaryRelation(
+                head=span, tail=span, label=self.loop_dummy_relation_name
+            )
+            encoded_relation = self.relation_encoder_decoder.encode(
+                annotation=dummy_relation, metadata=metadata
+            )
+            if encoded_relation is not None:
+                relation_encodings[dummy_relation] = encoded_relation
+
+        # sort relations by start indices of head and tail
+        sorted_relations = sorted(relation_encodings, key=cmp_to_key(cmp_src_rel))
+
+        # build target tokens
+        tgt_tokens = [self.bos_id]
+        for rel in sorted_relations:
+            encoded_relation = relation_encodings[rel]
+            tgt_tokens.extend(encoded_relation)
+        tgt_tokens.append(self.eos_id)
+
+        # sanity check
+        _, invalid = self.sanitize_sequence(tag_seq=tgt_tokens[1:])
+        if not all(v == 0 for k, v in invalid.items() if k not in self.ignore_error_types):
+            decoded, invalid = self.decode_annotations(
+                EncodingWithIdsAndOptionalCpmTag(tgt_tokens), metadata=metadata
+            )
+            logger.warning(f"invalid: {invalid}, decoded: {decoded}")
+
+        if self.max_target_length is not None and len(tgt_tokens) > self.max_target_length:
+            raise ValueError(
+                f"encoding length {len(tgt_tokens)} exceeds max_length {self.max_target_length}"
+            )
+
+        # build CPM tag
+        if self.create_constraints:
+            if metadata is None or "src_len" not in metadata:
+                raise Exception("metadata with 'src_len' is required to create constraints")
+            src_len = metadata["src_len"]
+            constraints = self.build_constraints(src_len=src_len, tgt_tokens=tgt_tokens)
+            cpm_tag = constraints
+        else:
+            cpm_tag = None
+        return EncodingWithIdsAndOptionalCpmTag(tgt_tokens=tgt_tokens, CPM_tag=cpm_tag)
+
+    def decode_annotations(
+        self, encoding: TaskOutputType, metadata: Optional[Dict[str, Any]] = None
+    ) -> Tuple[Dict[str, List[Annotation]], Any]:
+        # strip the bos token
+        ps, _errors = self.sanitize_sequence(tag_seq=encoding.tgt_tokens[1:])
+        relation_tuples: List[Tuple[Tuple[int, int], Tuple[int, int], str]] = []
+        entity_labels: Dict[Tuple[int, int], List[str]] = defaultdict(list)
+        for tup in ps:
+            rel = self.relation_encoder_decoder.decode(encoding=list(tup), metadata=metadata)
+            head_span = (rel.head.start, rel.head.end)
+            entity_labels[head_span].append(rel.head.label)
+
+            if rel.label != self.none_label:
+                tail_span = (rel.tail.start, rel.tail.end)
+                entity_labels[tail_span].append(rel.tail.label)
+                relation_tuples.append((head_span, tail_span, rel.label))
+            else:
+                assert rel.head == rel.tail
+
+        # TODO: do we really need this majority vote? Why not keep both? Or parametrize this?
+        entities: Dict[Tuple[int, int], LabeledSpan] = {}
+        for (start, end), labels in entity_labels.items():
+            c = Counter(labels)
+            # if len(c) > 1:
+            #    logger.warning(f"multiple labels for span, take the most common: {dict(c)}")
+            most_common_label = c.most_common(1)[0][0]
+            entities[(start, end)] = LabeledSpan(start=start, end=end, label=most_common_label)
+
+        entity_layer = list(entities.values())
+        relation_layer = [
+            BinaryRelation(head=entities[head_span], tail=entities[tail_span], label=label)
+            for head_span, tail_span, label in relation_tuples
+        ]
+        return {
+            self.span_layer_name: entity_layer,
+            self.relation_layer_name: relation_layer,
+        }, _errors
+
+    # TODO: use torch instead of numpy?
+    def _pointer_tag(
+        self,
+        last: List[int],
+        t: int,
+        idx: int,
+        arr: np.ndarray,
+    ) -> np.ndarray:
+        if t == 0:  # start # c1 [0, 1]
+            arr[: self.pointer_offset] = 0
+        elif idx % 7 == 0:  # c1 [0,1, 23]
+            arr[:t] = 0
+        elif idx % 7 == 1:  # tc1 [0,1,23, tc] span标签设为1
+            arr = np.zeros_like(arr, dtype=int)
+            for i in self.span_ids:
+                arr[i] = 1
+        elif idx % 7 == 2:  # c2 [0,1,23,tc, 45]
+            arr[: self.pointer_offset] = 0
+            arr[last[-3] : last[-2]] = 0
+        elif idx % 7 == 3:  # c2 [0,1,23,tc,45, 67]
+            arr[:t] = 0
+            if t < last[-4]:
+                arr[last[-4] :] = 0
+            else:
+                arr[last[-4] : last[-3]] = 0
+        elif idx % 7 == 4:  # tc2 [0,1,23,tc,45,67, tc]
+            arr = np.zeros_like(arr, dtype=int)
+            for i in self.span_ids:
+                arr[i] = 1
+        elif idx % 7 == 5:  # r [0,1,23,tc,45,67,tc, r]
+            arr = np.zeros_like(arr, dtype=int)
+            for i in self.relation_ids:
+                arr[i] = 1
+        elif idx % 7 == 6:  # next
+            arr[: self.pointer_offset] = 0
+        return arr
+
+    def build_constraints(
+        self,
+        src_len: int,
+        tgt_tokens: List[int],
+    ) -> List[List[int]]:
+        # strip the bos token
+        target = tgt_tokens[1:]
+        # pad for 0
+        likely_hood = np.ones(src_len + self.pointer_offset, dtype=int)
+        likely_hood[: self.pointer_offset] = 0
+        CMP_tag: List[np.ndarray] = [likely_hood]
+        for idx, t in enumerate(target[:-1]):
+            last7 = target[idx - 7 if idx - 7 > 0 else 0 : idx + 1]
+            likely_hood = np.ones(src_len + self.pointer_offset, dtype=int)
+            tag = self._pointer_tag(last=last7, t=t, idx=idx, arr=likely_hood)
+            tag[self.none_id] = 1
+            CMP_tag.append(tag)
+        last_end = np.zeros(src_len + self.pointer_offset, dtype=int)
+        last_end[self.none_id] = 1
+        last_end[target[-1]] = 1
+        CMP_tag[-1] = last_end
+        result = [i.tolist() for i in CMP_tag]
+        return result
+
     def maybe_log_example(
         self,
         task_encoding: TaskEncodingType,
@@ -261,11 +640,11 @@ class PointerNetworkTaskModule(
             src_tokens = self.tokenizer.convert_ids_to_tokens(src_token_ids)
             tgt_token_ids = targets.tgt_tokens
             tgt_tokens = [
-                self.annotation_encoder_decoder.targets[tgt_token_id]
-                if tgt_token_id < self.annotation_encoder_decoder.pointer_offset
+                self.targets[tgt_token_id]
+                if tgt_token_id < self.pointer_offset
                 else str(tgt_token_id)
                 + " {"
-                + str(src_tokens[tgt_token_id - self.annotation_encoder_decoder.pointer_offset])
+                + str(src_tokens[tgt_token_id - self.pointer_offset])
                 + "}"
                 for tgt_token_id in tgt_token_ids
             ]
@@ -327,9 +706,9 @@ class PointerNetworkTaskModule(
 
         layers = {
             layer_name: self.get_mapped_layer(document, layer_name=layer_name)
-            for layer_name in self.annotation_encoder_decoder.layer_names
+            for layer_name in self.layer_names
         }
-        result = self.annotation_encoder_decoder.encode(
+        result = self.encode_annotations(
             layers=layers,
             metadata={**task_encoding.metadata, "src_len": task_encoding.inputs.src_seq_len},
         )
@@ -356,15 +735,22 @@ class PointerNetworkTaskModule(
 
         return inputs, targets
 
-    def unbatch_output(self, model_output: ModelBatchOutput) -> Sequence[TaskOutput]:
-        return self.annotation_encoder_decoder.unbatch(prediction=model_output)
+    def unbatch_output(self, model_output: ModelBatchOutput) -> Sequence[TaskOutputType]:
+        # model_output just contains "pred"
+        pred = model_output["pred"]
+        batch_size = pred.size(0)
+        result = [
+            EncodingWithIdsAndOptionalCpmTag(pred[i].to(device="cpu").tolist())
+            for i in range(batch_size)
+        ]
+        return result
 
     def create_annotations_from_output(
         self,
         task_encoding: TaskEncodingType,
-        task_output: TaskOutput,
+        task_output: TaskOutputType,
     ) -> Iterator[Tuple[str, Annotation]]:
-        layers, _errors = self.annotation_encoder_decoder.decode(
+        layers, _errors = self.decode_annotations(
             encoding=task_output, metadata=task_encoding.metadata
         )
         tokenized_document = task_encoding.metadata["tokenized_document"]
