+++ tests/models/test_simple_pointer_network.py
@@ -69,15 +69,15 @@ def document():
 @pytest.fixture(scope="module")
 def taskmodule(document):
     taskmodule = PointerNetworkTaskModule(
-        annotation_encoder_decoder_kwargs={
-            "span_layer_name": "entities",
-            "relation_layer_name": "relations",
-            "exclude_labels_per_layer": {"relations": ["no_relation"]},
-        },
+        span_layer_name="entities",
+        relation_layer_name="relations",
+        exclude_labels_per_layer={"relations": ["no_relation"]},
         annotation_field_mapping={
             "entities": "labeled_spans",
             "relations": "binary_relations",
         },
+        create_constraints=False,
+        # tokenizer_kwargs={"strict_span_conversion": False},
     )
 
     taskmodule.prepare(documents=[document])
@@ -86,71 +86,7 @@ def taskmodule(document):
 
 
 def test_taskmodule(taskmodule):
-    # check the annotation_encoder_decoder
-    annotation_encoder_decoder = taskmodule.annotation_encoder_decoder
-    assert annotation_encoder_decoder.is_prepared
-    assert annotation_encoder_decoder.prepared_attributes == {
-        "labels_per_layer": {
-            "entities": ["content", "person", "topic"],
-            "relations": ["is_about"],
-        },
-    }
-    assert annotation_encoder_decoder.layer_names == ["entities", "relations"]
-    assert annotation_encoder_decoder.special_targets == ["<s>", "</s>"]
-    assert annotation_encoder_decoder.labels == ["none", "content", "person", "topic", "is_about"]
-    assert annotation_encoder_decoder.targets == [
-        "<s>",
-        "</s>",
-        "none",
-        "content",
-        "person",
-        "topic",
-        "is_about",
-    ]
-    assert annotation_encoder_decoder.bos_id == 0
-    assert annotation_encoder_decoder.eos_id == 1
-    assert annotation_encoder_decoder.none_id == 2
-    assert annotation_encoder_decoder.span_ids == [3, 4, 5]
-    assert annotation_encoder_decoder.relation_ids == [6]
-    assert annotation_encoder_decoder.label2id == {
-        "content": 3,
-        "is_about": 6,
-        "none": 2,
-        "person": 4,
-        "topic": 5,
-    }
-
-    # check taskmodule properties
-    assert taskmodule.prepared_attributes == {
-        "annotation_encoder_decoder_kwargs": {
-            "span_layer_name": "entities",
-            "relation_layer_name": "relations",
-            "exclude_labels_per_layer": {"relations": ["no_relation"]},
-            "bos_token": "<s>",
-            "eos_token": "</s>",
-            "labels_per_layer": {
-                "entities": ["content", "person", "topic"],
-                "relations": ["is_about"],
-            },
-        }
-    }
-    assert taskmodule.label_embedding_weight_mapping == {
-        50265: [45260],
-        50266: [39763],
-        50267: [354, 1215, 9006],
-        50268: [5970],
-        50269: [10166],
-    }
-    assert taskmodule.target_tokens == [
-        "<s>",
-        "</s>",
-        "<<none>>",
-        "<<content>>",
-        "<<person>>",
-        "<<topic>>",
-        "<<is_about>>",
-    ]
-    assert taskmodule.target_token_ids == [0, 2, 50266, 50269, 50268, 50265, 50267]
+    assert taskmodule.is_prepared
 
 
 @pytest.fixture(scope="module")
@@ -158,11 +94,15 @@ def model(taskmodule, config) -> SimplePointerNetworkModel:
     torch.manual_seed(42)
     model = SimplePointerNetworkModel(
         model_name_or_path="sshleifer/distilbart-xsum-12-1",
+        bos_id=taskmodule.bos_id,
+        eos_id=taskmodule.eos_id,
+        pad_id=taskmodule.eos_id,
+        label_ids=taskmodule.label_ids,
         target_token_ids=taskmodule.target_token_ids,
         vocab_size=len(taskmodule.tokenizer),
         embedding_weight_mapping=taskmodule.label_embedding_weight_mapping,
-        annotation_encoder_decoder_name=taskmodule.annotation_encoder_decoder_name,
-        annotation_encoder_decoder_kwargs=taskmodule.annotation_encoder_decoder_kwargs,
+        generation_kwargs=taskmodule.generation_kwargs,
+        taskmodule_config=taskmodule._config(),
         **config,
     )
     # set model to training mode, otherwise model.encoder.bart_encoder.training will be False!
@@ -307,11 +247,16 @@ def test_configure_optimizers_with_warmup_proportion(taskmodule, config):
     torch.manual_seed(42)
     model = SimplePointerNetworkModel(
         model_name_or_path="sshleifer/distilbart-xsum-12-1",
+        bos_id=taskmodule.bos_id,
+        eos_id=taskmodule.eos_id,
+        pad_id=taskmodule.eos_id,
+        label_ids=taskmodule.label_ids,
         target_token_ids=taskmodule.target_token_ids,
         vocab_size=len(taskmodule.tokenizer),
         embedding_weight_mapping=taskmodule.label_embedding_weight_mapping,
-        annotation_encoder_decoder_name=taskmodule.annotation_encoder_decoder_name,
-        annotation_encoder_decoder_kwargs=taskmodule.annotation_encoder_decoder_kwargs,
+        # annotation_encoder_decoder_name=taskmodule.annotation_encoder_decoder_name,
+        # annotation_encoder_decoder_kwargs=taskmodule.annotation_encoder_decoder_kwargs,
+        taskmodule_config=taskmodule._config(),
         warmup_proportion=0.1,
     )
     # set model to training mode, otherwise model.encoder.bart_encoder.training will be False!
